refactor(maintenance): extract archive rollup support modules
